Create Cycle_LinkedList.cpp
There is a cycle in a linked list if there is some node in the list that can be reached again by continuously following the next pointer. Internally, "pos" is used to denote the index of the node that tail's next pointer is connected to. The code returns true if there is a cycle in the linked list , else false